+++ lib/theme/app_theme_name.dart
@@ -1,0 +1,9 @@
+enum AppThemeName {
+  mint,
+  ice,
+  indigo,
+  sand,
+  lime;
+
+  static AppThemeName get defaultTheme => AppThemeName.mint;
+}